% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ghactions_ogd.R
\name{use_ghactions_ogd}
\alias{use_ghactions_ogd}
\title{Workflow automation with GitHub Actions (specialised)}
\usage{
use_ghactions_ogd(workflow)
}
\arguments{
\item{workflow}{see \code{\link[ghactions:use_ghactions]{ghactions::use_ghactions()}}}
}
\description{
See \code{\link[ghactions:use_ghactions]{use_ghactions()}} from the \link{ghactions} package
}
